Lint with type information and no escape hatches
Issue #20's decision, installed: ESLint 10.10.0 flat config, typescript-eslint
8.70.0 with `projectService: true`, `noInlineConfig`, and the rules living in a
workspace package rather than loose files.

Type-awareness is the whole reason for choosing ESLint here, so it is verified
rather than assumed. A file containing an unawaited promise reports
`@typescript-eslint/no-floating-promises`; the same file under a config with
`projectService: false` throws "you have used a rule which requires type
information" instead. The wiring is load-bearing, not decorative.

`noInlineConfig` is verified the same way. An
`eslint-disable-next-line @typescript-eslint/no-unnecessary-type-assertion`
placed over a violating line silences it with the option off, and with it on
the error reports anyway plus a warning that the directive has no effect. The
three inline suppressions the repo already carried (all naming
`react-hooks/exhaustive-deps`, which no installed plugin even provides) are
removed; the reason each carried is kept as prose, and two that had none now
have one.

`packages/eslint-rules` is a skeleton on purpose: an empty rule set, wired into
the flat config so that loading it is proven. `enq` purity (#22) and
no-styles-outside-`packages/ui` (#12) are deliberately not stubbed, because a
stub that always passes is indistinguishable from a rule that has stopped
firing. The package is consumed as source — its export map points at
`src/index.ts`, Node 26 strips the types on load, and `erasableSyntaxOnly` in
its tsconfig fails the typecheck on anything Node could not simply erase — so
`eslint.config.js` needs no build step ahead of it.

Exemptions are flat-config `files:` globs, never comments. `eslint --print-config`
confirms the `map-editor` plugin is attached to `src/core/document.ts` and to a
hypothetical `packages/fixtures/src/sample.ts`, and absent from
`src/core/core.test.ts`, `src/core/mesher/terrain.bench.ts`, `scripts/probe.mjs`
and `packages/fixtures/src/sample.test.ts`. That is the split the issue's final
correction asks for: tests and `scripts/` describe the system from outside it,
`fixtures` produces data that flows into it.

The rule set is `js.configs.recommended` plus `recommendedTypeChecked` in full.
It found 23 violations across 7,596 lines of TypeScript, which is not a flood.
Twenty-one of them are fixed rather than silenced, and all but one of those are
type-only edits the compiler erases, so the production risk is close to nil:

- 7 `no-unnecessary-type-assertion` (autofixed; a dead `type Edit` import fell
  out of the store.ts one)
- 5 `no-unsafe-assignment` from `new Array(n)` being `any[]` in `createMap`, and
  from `JSON.parse` in `deserialize` — the terrain block was entering the
  document untyped
- 6 `no-unsafe-assignment`/`no-unsafe-member-access` in the io tests, now going
  through one named `OnDisk` shape instead of three untyped `JSON.parse` calls
- 2 `no-redundant-type-constituents`: `Direction | number` collapses to `number`,
  so it read as a constraint while enforcing nothing
- 1 `no-misused-promises`: an async handler on a `<button onClick>`, now `void`-
  wrapped, which is the idiom the same file already uses five lines above

Two rules are off, named individually with their reasons in the config. Both
land on `buildExportScene`/`exportGltf` in `src/runtime/export.ts`, which no
test covers: `require-await` wants its exported return type changed from
`Promise<Scene>` to `Scene`, and `prefer-promise-reject-errors` wants the glTF
exporter's error wrapped, changing what its one caller catches. Reshaping
untested exported surface is not a toolchain commit's call.

`vitest.config.ts` joins tsconfig's `include` — the project service cannot lint
a file no tsconfig claims, and this also brings it under `tsc --noEmit`.

`pnpm lint` runs `eslint .` from the root, which covers `packages/*` too, and
`//#lint` joins the turbo pipeline alongside `gate`/`gate:full`. Verified the
gate actually goes red through turbo: appending an unawaited call to
`src/core/chunks.ts` turned `//:lint` from a cache hit into
`cache miss, executing c5ecb50e346c188d` and failed the run with exit 1.

Gate: 4 tasks successful, 53/53 tests, boundaries OK, exit 0.

+++ src/runtime/export.ts
@@ -153,10 +153,10 @@ export async function buildExportScene(doc: MapDoc, options: ExportOptions): Pro
     let indexOffset = 0
     for (const geometry of mergedPositions) {
       const position = geometry.getAttribute('position')
-      positions.set(position.array as Float32Array, vertexOffset * 3)
-      normals.set(geometry.getAttribute('normal').array as Float32Array, vertexOffset * 3)
-      uvs.set(geometry.getAttribute('uv').array as Float32Array, vertexOffset * 2)
-      colors.set(geometry.getAttribute('color').array as Float32Array, vertexOffset * 3)
+      positions.set(position.array, vertexOffset * 3)
+      normals.set(geometry.getAttribute('normal').array, vertexOffset * 3)
+      uvs.set(geometry.getAttribute('uv').array, vertexOffset * 2)
+      colors.set(geometry.getAttribute('color').array, vertexOffset * 3)
       const index = geometry.getIndex()
       if (index) {
         for (let i = 0; i < index.count; i++) indices[indexOffset + i] = index.getX(i) + vertexOffset
